requirements.txt
django==5.0.1
django-cors-headers>=4.2.0,<=4.3.1
djangorestframework>=3.13.0,<3.15.0
django-rest-swagger>=2.2.0,<2.3.0
drf-yasg==1.21.4
djoser==2.2.0